Compose command callbacks
Previously, some commands and all callers of push_layer would discard
any existing callback. The callback can already exist if running in a
sequence of commands.

Since macros use the callback, this solves edge-cases using macros in
key-bind sequences where a macro would clash with another command which
uses the callback field. For example two macros would cause this:

    C = ["@t<ret>", "@c"]

Would only execute the `c` key prior to this change.

Also, any command that pushes a layer to the compositor or otherwise
set the `cx.callback` field would have that effect ignored:

    ["shell_append_output", "@echo foo<ret>"]

`shell_append_output` pushes a compositor layer for the prompt which
sets the `cx.callback` field. Prior to this change, the creation of the
prompt was ignored so instead the macro was executed within the current
mode rather than in the prompt.