Use macro text for naming and docs
This resolves the new `unimplemented!` calls introduced with macro
key-binds. The one in `keymap.rs` is necessary because the command
palette will attempt to construct a reverse mapping from the current
keymap, so the `unimplemented!` will panic on opening the command
palette.

The name and doc text for these can be the unparsed macro text when
the macro is implemented in a keymap. If any commands are implemented
via macros in the future, we could use custom values for fields.